Simple turns are working
Turns happen in phases. Right now we're skipping the Start phase, but
it could be used for animations and such to show who's turn it is.

Game simply holds a Vec of all active Skirmers, which it moves through
when the current turn Skirmer gets to TurnPhase::Finish.

Initial TODO:
    * update GUI to show how much a unit can move
    * update GUI to show distance from each unit during their turn
    * end turn GUI element